requirements.txt
Django==1.8.3
Django-Select2==5.8.9
PbPython==0.1.9
PyYAML==3.12
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
cffi==1.8.3
cryptography==1.5.2
django-appconf==1.0.2
django-chosen==0.1
django-compressor==2.1
django-debug-toolbar==1.5
django-extensions==1.7.4
django-fontawesome==0.2.6
django-multiupload==0.5.2
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
kombu==3.0.37
ndg-httpsclient==0.4.2
oauthlib==2.0.0
pbr==1.10.0
pyOpenSSL==16.1.0
pyasn1==0.1.9
pycparser==2.14
pytz==2016.7
rcssmin==1.0.6
requests==2.11.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.1
stevedore==1.17.1
tweepy==3.5.0
urllib3==1.18
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wsgiref==0.1.2
